bring up to date

+++ src/parameters.h
@@ -11,9 +11,9 @@ class Parameters
     int thin;
     int burnin;
     int samples;
-    int complexity_limit;
-    int importance_sampling_depth;
-    double importance_sampling_scaling_factor;
+    long int complexity_limit;
+    long int importance_sampling_depth;
+    long int importance_sampling_scaling_factor;
 
     // Model Parameters
     // Complexity of Infection
